Cleanup / refactor
